refactor FHIRFunctions azure aspect, add comments
