+++ prime-router/src/main/kotlin/azure/ActionHistory.kt
@@ -342,7 +342,10 @@ class ActionHistory : Logging {
         reportFile.itemCount = report.itemCount
         reportsReceived[reportFile.reportId] = reportFile
 
-        // add to queue
+        // batch queue messages are added by the batchDecider, not ActionHistory.
+        // TODO: Need to update this process to have a better way to determine what messages should be sent
+        //  automatically as part of queueMessages and what are being send manually as part of the parent function.
+        //  The automatic queueing uses the action name as the queue name, and this is not the case for FHIR actions
         if (event.eventAction != Event.EventAction.BATCH &&
             event.eventAction != Event.EventAction.ROUTE &&
             event.eventAction != Event.EventAction.TRANSLATE
@@ -404,6 +407,9 @@ class ActionHistory : Logging {
         trackItemLineages(report)
 
         // batch queue messages are added by the batchDecider, not ActionHistory
+        // TODO: Need to update this process to have a better way to determine what messages should be sent
+        //  automatically as part of queueMessages and what are being send manually as part of the parent function.
+        //  The automatic queueing uses the action name as the queue name, and this is not the case for FHIR actions
         if (event.eventAction != Event.EventAction.BATCH &&
             event.eventAction != Event.EventAction.ROUTE &&
             event.eventAction != Event.EventAction.TRANSLATE
@@ -436,7 +442,8 @@ class ActionHistory : Logging {
 
         // batch queue messages are added by the batchDecider, not ActionHistory
         // TODO: Need to update this process to have a better way to determine what messages should be sent
-        //  automatically as part of queueMessages and what are being send manually as part of the parent function
+        //  automatically as part of queueMessages and what are being send manually as part of the parent function.
+        //  The automatic queueing uses the action name as the queue name, and this is not the case for FHIR actions
         if (event.eventAction != Event.EventAction.BATCH &&
             event.eventAction != Event.EventAction.ROUTE &&
             event.eventAction != Event.EventAction.TRANSLATE
